Move base getter back to base module
