+++ luchador/nn/util/getter.py
@@ -3,110 +3,16 @@ from __future__ import absolute_import
 
 from luchador.util import get_subclasses
 import luchador.nn
-from ..base import BaseCost, BaseLayer, BaseOptimizer, BaseInitializer
 from ..base import wrapper as base_wrapper
+from ..base.getter import get_cost, get_layer, get_optimizer, get_initializer
 from ..model import BaseModel
 
-
-###############################################################################
-# Functions for fetching class
-def get_cost(name):
-    """Get ``Cost`` class by name
-
-    Parameters
-    ----------
-    name : str
-        Type of ``Cost`` to get.
-
-    Returns
-    -------
-    type
-        ``Cost`` type found
-
-    Raises
-    ------
-    ValueError
-        When ``Cost`` class with the given type is not found
-    """
-    for class_ in get_subclasses(BaseCost):
-        if class_.__name__ == name:
-            return class_
-    raise ValueError('Unknown Cost: {}'.format(name))
+__all__ = [
+    'get_cost', 'get_layer', 'get_optimizer', 'get_initializer',
+    'get_model', 'get_input', 'get_tensor', 'get_grad', 'get_operation',
+]
 
 
-def get_layer(name):
-    """Get ``Layer`` class by name
-
-    Parameters
-    ----------
-    name : str
-        Type of ``Layer`` to get
-
-    Returns
-    -------
-    type
-        ``Layer`` type found
-
-    Raises
-    ------
-    ValueError
-        When ``Layer`` class with the given type is not found
-    """
-    for class_ in get_subclasses(BaseLayer):
-        if class_.__name__ == name:
-            return class_
-    raise ValueError('Unknown Layer: {}'.format(name))
-
-
-def get_optimizer(name):
-    """Get ``Optimizer`` class by name
-
-    Parameters
-    ----------
-    name : str
-        Type of ``Optimizer`` to get
-
-    Returns
-    -------
-    type
-        ``Optimizer`` type found
-
-    Raises
-    ------
-    ValueError
-        When ``Optimizer`` class with the given type is not found
-    """
-    for class_ in get_subclasses(BaseOptimizer):
-        if class_.__name__ == name:
-            return class_
-    raise ValueError('Unknown Optimizer: {}'.format(name))
-
-
-def get_initializer(name):
-    """Get ``Initializer`` class by name
-
-    Parameters
-    ----------
-    name : str
-        Type of ``Initializer`` to get
-
-    Returns
-    -------
-    type
-        ``Initializer`` type found
-
-    Raises
-    ------
-    ValueError
-        When ``Initializer`` class with the given type is not found
-    """
-    for class_ in get_subclasses(BaseInitializer):
-        if class_.__name__ == name:
-            return class_
-    raise ValueError('Unknown Initializer: {}'.format(name))
-
-
-# -----------------------------------------------------------------------------
 def get_model(name):
     """Get ``Model`` class by name
 
@@ -125,7 +31,7 @@ def get_model(name):
     ValueError
         When ``Model`` class with the given name is not found
     """
-    for class_ in luchador.util.get_subclasses(BaseModel):
+    for class_ in get_subclasses(BaseModel):
         if class_.__name__ == name:
             return class_
     raise ValueError('Unknown model: {}'.format(name))
